+++ app/src/main/java/com/example/friendlygram/MainActivity.kt
@@ -9,6 +9,7 @@ import androidx.appcompat.widget.Toolbar
 import com.example.friendlygram.databinding.ActivityMainBinding
 import com.example.friendlygram.ui.ChatFragment
 import com.example.friendlygram.ui.SettingsFragment
+import com.example.friendlygram.ui.objects.AppDrawer
 import com.mikepenz.materialdrawer.AccountHeader
 import com.mikepenz.materialdrawer.AccountHeaderBuilder
 import com.mikepenz.materialdrawer.Drawer
@@ -22,8 +23,9 @@ import java.util.function.Consumer
 class MainActivity : AppCompatActivity() {
 
     private lateinit var mBinding: ActivityMainBinding
-    private lateinit var mDrawer: Drawer
-    private lateinit var mHeader: AccountHeader
+
+    private lateinit var mAppDrawer: AppDrawer
+
     private lateinit var mToolbar: Toolbar
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -41,101 +43,16 @@ class MainActivity : AppCompatActivity() {
 
     private fun initFunc() {
         setSupportActionBar(mToolbar)
+        mAppDrawer.create()
         supportFragmentManager.beginTransaction()
-            .replace(R.id.dataContainer,ChatFragment()).commit()
-        createHeader()
-        createDrawer()
+            .replace(R.id.dataContainer, ChatFragment()).commit()
+
     }
 
-    private fun createDrawer() {
-        mDrawer = DrawerBuilder()
-            .withActivity(this)
-            .withToolbar(mToolbar)
-            .withActionBarDrawerToggle(true)
-            .withSelectedItem(-1)
-            .withAccountHeader(mHeader)
-            .addDrawerItems(
-                PrimaryDrawerItem().withIdentifier(100)
-                    .withIconTintingEnabled(true)
-                    .withName("Создать группу")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_create_groups),
-                PrimaryDrawerItem().withIdentifier(101)
-                    .withIconTintingEnabled(true)
-                    .withName("Создать секретный чат")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_secret_chat),
-                PrimaryDrawerItem().withIdentifier(102)
-                    .withIconTintingEnabled(true)
-                    .withName("Создать канал")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_create_channel),
-                PrimaryDrawerItem().withIdentifier(103)
-                    .withIconTintingEnabled(true)
-                    .withName("Контакты")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_contacts),
-                PrimaryDrawerItem().withIdentifier(104)
-                    .withIconTintingEnabled(true)
-                    .withName("Звонки")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_phone),
-                PrimaryDrawerItem().withIdentifier(105)
-                    .withIconTintingEnabled(true)
-                    .withName("Избранное")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_favorites),
-                PrimaryDrawerItem().withIdentifier(106)
-                    .withIconTintingEnabled(true)
-                    .withName("Настройки")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_settings),
-                DividerDrawerItem(),
-                PrimaryDrawerItem().withIdentifier(107)
-                    .withIconTintingEnabled(true)
-                    .withName("Пригласить друзей")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_invate),
-                PrimaryDrawerItem().withIdentifier(108)
-                    .withIconTintingEnabled(true)
-                    .withName("Вопросы о Friendlygram")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_help),
 
-
-
-
-            ).withOnDrawerItemClickListener(object :Drawer.OnDrawerItemClickListener{
-                override fun onItemClick(
-                    view: View?,
-                    position: Int,
-                    drawerItem: IDrawerItem<*>
-                ): Boolean {
-                    when (position) {
-                        7 -> supportFragmentManager.beginTransaction()
-                            .addToBackStack(null)
-                            .replace(R.id.dataContainer,SettingsFragment()).commit()
-                    }
-                    return false
-                }
-
-
-
-            }).build()
-    }
-
-    private fun createHeader() {
-mHeader = AccountHeaderBuilder()
-    .withActivity(this)
-    .withHeaderBackground(R.drawable.header)
-    .addProfiles(
-        ProfileDrawerItem().withName("COINGEN")
-            .withEmail("+79111111111")
-    ).build()
-    }
-
-    private fun initFields(){
+    private fun initFields() {
         mToolbar = mBinding.mainToolbar
+        mAppDrawer = AppDrawer(this,mToolbar)
 
     }
 }
